Implement getParent API.
getParent which tries to get the Parent object
considering the "name" argument as the corporate
name. If no parent is found then the name is
considered as a parent alias and an attempt is made
to get check if the alias is a unique alias and if so then
the corresponding parent is returned.